requirements.txt
face_recognition_models
Click>=6.0
dlib>=19.3.0
numpy
Pillow
scipy>=0.17.0

# FaceRec Pro dashboard (FaceRecPro/app.py)
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
opencv-python-headless>=4.8.0
mysql-connector-python>=8.0.0
